Update: Dockerfileも変更するように

+++ build_util/build_preprocess.bash
@@ -4,6 +4,10 @@ if [ ! -v DOWNLOAD_RESOURCE_PATH ]; then
     echo "DOWNLOAD_RESOURCE_PATHが未定義です"
     exit 1
 fi
+if [ ! -v VERSION ]; then
+    echo "VERSIONが未定義です"
+    exit 1
+fi
 
 rm -r speaker_info
 cp -r $DOWNLOAD_RESOURCE_PATH/character_info speaker_info
@@ -14,7 +18,10 @@ rm speaker_info/*/icons/*.png_large
 # マニフェスト
 jq -s '.[0] * .[1]' engine_manifest.json $DOWNLOAD_RESOURCE_PATH/engine/engine_manifest.json \
     > engine_manifest.json.tmp
-mv engine_manifest.json.tmp engine_manifest.json
+jq '.version = "$VERSION"' engine_manifest.json.tmp \
+    > engine_manifest.json
+rm engine_manifest.json.tmp
+
 
 python build_util/merge_update_infos.py \
     engine_manifest_assets/update_infos.json \
